cat_unlearn_2_cat/requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
scipy==1.13.1
seaborn==0.13.2
pingouin==0.5.4
Pillow==10.3.0
pygame==2.5.2
